+++ packages/monaco/src/browser/contrib/merge-editor/merge-editor.service.ts
@@ -75,7 +75,5 @@ export class MergeEditorService extends Disposable {
     const { changes: changes2 } = result2;
     this.resultView.inputDiffComputingResult(changes2, 0);
     this.incomingView.inputDiffComputingResult(changes2);
-
-    this.currentView.layout();
   }
 }
