+++ packages/monaco/src/browser/contrib/merge-editor/view/editors/currentCodeEditor.tsx
@@ -1,6 +1,7 @@
 import { Injectable, Autowired, Injector, INJECTOR_TOKEN } from '@opensumi/di';
-import { MonacoService } from '@opensumi/ide-core-browser';
+import { Event, MonacoService } from '@opensumi/ide-core-browser';
 import { ICodeEditor } from '@opensumi/monaco-editor-core/esm/vs/editor/browser/editorBrowser';
+import { EditorLayoutInfo } from '@opensumi/monaco-editor-core/esm/vs/editor/common/config/editorOptions';
 import { Range } from '@opensumi/monaco-editor-core/esm/vs/editor/common/core/range';
 import { LineRange, LineRangeMapping } from '@opensumi/monaco-editor-core/esm/vs/editor/common/diff/linesDiffComputer';
 import { IStandaloneEditorConstructionOptions } from '@opensumi/monaco-editor-core/esm/vs/editor/standalone/browser/standaloneCodeEditor';
@@ -19,7 +20,7 @@ import { BaseCodeEditor } from './baseCodeEditor';
 @Injectable({ multiple: false })
 export class CurrentCodeEditor extends BaseCodeEditor {
   protected getMonacoEditorOptions(): IStandaloneEditorConstructionOptions {
-    return { lineNumbersMinChars: 2 };
+    return { readOnly: true };
   }
 
   private rangeMapping: LineRangeMapping[] = [];
@@ -65,28 +66,6 @@ export class CurrentCodeEditor extends BaseCodeEditor {
 
   protected getRetainLineWidget(): GuidelineWidget[] {
     return [];
-  }
-
-  public override mount(): void {
-    super.mount();
-
-    const marginWith = this.editor.getLayoutInfo().contentLeft;
-
-    this.addDispose(
-      this.decorations.onDidChangeDecorations((decorations: MergeEditorDecorations) => {
-        const widgets = decorations.getLineWidgets();
-        if (widgets.length > 0) {
-          widgets.forEach((w) => {
-            if (w) {
-              w.setContainerStyle({
-                width: `calc(100% - ${marginWith}px)`,
-              });
-            }
-          });
-        }
-        this.layout();
-      }),
-    );
   }
 
   public inputDiffComputingResult(changes: LineRangeMapping[]): void {
